Added team data for selection
Added all the avaiable teams for the selection, managing also the fact that when a team (A) is selected, the same team should not appear in the second list, but also, should reapear when team A change again.
The result is shown in a separate <div> element on the main screen. Later, will be deleted.